Revert "Sif dump update"
