+++ indra_db/__init__.py
@@ -1,2 +1,2 @@
-from .util import get_primary_db, get_db, get_ro
 from .databases import texttypes, formats, sql_expressions
+from .util import get_primary_db, get_db
