+++ indra_db/util/__init__.py
@@ -9,10 +9,9 @@ Some key functions' capabilities include:
 __all__ = ['get_primary_db', 'get_db', 'insert_raw_agents', 'insert_pa_stmts',
            'insert_pa_agents', 'insert_db_stmts', 'get_raw_stmts_frm_db_list',
            'distill_stmts', 'regularize_agent_id', 'get_statement_object',
-           'extract_agent_data', 'get_ro', 'S3Path']
+           'extract_agent_data', 'get_ro']
 
 from .insert import *
-from .s3_path import *
 from .helpers import *
 from .constructors import *
 from .content_scripts import *
