+++ indra_db/util/dump_sif.py
@@ -1,15 +1,10 @@
-__all__ = ['load_db_content', 'make_dataframe', 'get_source_counts', 'NS_LIST',
-           'dump_sif']
+__all__ = ['load_db_content', 'make_dataframe', 'make_ev_strata', 'NS_LIST']
 
 import pickle
 import logging
 import argparse
-from io import StringIO
-from datetime import datetime
 from itertools import permutations
 from collections import OrderedDict
-
-from indra.util.aws import get_s3_client
 
 try:
     import pandas as pd
@@ -17,12 +12,9 @@ except ImportError:
     print("Pandas not available.")
     pd = None
 
-from indra_db.util.s3_path import S3Path
-from indra_db.util.constructors import get_ro, get_db
+from indra_db import util as dbu
 
 logger = logging.getLogger(__name__)
-S3_SIF_BUCKET = 'bigmech'
-S3_SUBDIR = 'indra_db_sif_dump'
 
 NS_PRIORITY_LIST = (
     'FPLX',
@@ -42,81 +34,40 @@ NS_LIST = ('NAME', 'MIRBASE', 'HGNC', 'FPLX', 'GO', 'MESH', 'HMDB', 'CHEBI',
            'PUBCHEM')
 
 
-def _pseudo_key(fname, ymd_date):
-    return S3Path.from_key_parts(S3_SIF_BUCKET, S3_SUBDIR, ymd_date, fname)
-
-
-def upload_pickle_to_s3(obj, s3_path):
-    """Upload a python object as a pickle to s3"""
-    logger.info('Uploading %s as pickle object to bucket %s'
-                % (s3_path.key.split('/')[-1], s3_path.bucket))
-    s3 = get_s3_client(unsigned=False)
-    try:
-        s3.put_object(Body=pickle.dumps(obj=obj), **s3_path.kw())
-        logger.info('Finished dumping file to s3')
-    except Exception as e:
-        logger.error('Failed to upload to s3')
-        logger.exception(e)
-
-
-def load_pickle_from_s3(s3_path):
-    logger.info('Loading pickle %s.' % s3_path)
-    s3 = get_s3_client(unsigned=False)
-    try:
-        res = s3.get_object(**s3_path.kw())
-        obj = pickle.loads(res['Body'].read())
-        logger.info('Finished loading %s.' % s3_path)
-        return obj
-    except Exception as e:
-        logger.error('Failed to load %s.' % s3_path)
-        logger.exception(e)
-
-
-def load_db_content(ns_list, pkl_filename=None, ro=None, reload=False):
-    if isinstance(pkl_filename, str) and pkl_filename.startswith('s3:'):
-        pkl_filename = S3Path.from_string(pkl_filename)
+def load_db_content(reload, ns_list, pkl_filename=None, db=None):
     # Get the raw data
-    if reload or not pkl_filename:
-        if not ro:
-            ro = get_ro('primary')
+    if reload:
+        if not db:
+            db = dbu.get_primary_db()
         logger.info("Querying the database for statement metadata...")
         results = []
         for ns in ns_list:
             logger.info("Querying for {ns}".format(ns=ns))
-            res = ro.select_all([ro.PaMeta.mk_hash, ro.PaMeta.db_name,
-                                 ro.PaMeta.db_id, ro.PaMeta.ag_num,
-                                 ro.PaMeta.ev_count, ro.PaMeta.type],
-                                ro.PaMeta.db_name.like(ns))
+            res = db.select_all([db.PaMeta.mk_hash, db.PaMeta.db_name,
+                                 db.PaMeta.db_id, db.PaMeta.ag_num,
+                                 db.PaMeta.ev_count, db.PaMeta.type],
+                                db.PaMeta.db_name.like(ns))
             results.extend(res)
         results = set(results)
         if pkl_filename:
-            if isinstance(pkl_filename, S3Path):
-                upload_pickle_to_s3(results, pkl_filename)
-            else:
-                with open(pkl_filename, 'wb') as f:
-                    pickle.dump(results, f)
-    # Get a cached pickle
-    else:
+            with open(pkl_filename, 'wb') as f:
+                pickle.dump(results, f)
+    elif pkl_filename:
         logger.info("Loading database content from %s" % pkl_filename)
-        if pkl_filename.startswith('s3:'):
-            results = load_pickle_from_s3(pkl_filename)
-        else:
-            with open(pkl_filename, 'rb') as f:
-                results = pickle.load(f)
+        with open(pkl_filename, 'rb') as f:
+            results = pickle.load(f)
     logger.info("{len} stmts loaded".format(len=len(results)))
     return results
 
 
-def get_source_counts(pkl_filename=None, ro=None):
+def make_ev_strata(pkl_filename=None, ro=None):
     """Returns a dict of dicts with evidence count per source, per statement
 
     The dictionary is at the top level keyed by statement hash and each
     entry contains a dictionary keyed by the source that support the
     statement where the entries are the evidence count for that source."""
-    if isinstance(pkl_filename, str) and pkl_filename.startswith('s3:'):
-        pkl_filename = S3Path.from_string(pkl_filename)
     if not ro:
-        ro = get_ro('primary-ro')
+        ro = dbu.get_ro('primary')
     res = ro.select_all(ro.PaStmtSrc)
     ev = {}
     for r in res:
@@ -126,17 +77,12 @@ def get_source_counts(pkl_filename=None, ro=None):
                              rd[k] is not None}
 
     if pkl_filename:
-        if isinstance(pkl_filename, S3Path):
-            upload_pickle_to_s3(obj=ev, s3_path=pkl_filename)
-        else:
-            with open(pkl_filename, 'wb') as f:
-                pickle.dump(ev, f)
+        with open(pkl_filename, 'wb') as f:
+            pickle.dump(ev, f)
     return ev
 
 
 def make_dataframe(reconvert, db_content, pkl_filename=None):
-    if isinstance(pkl_filename, str) and pkl_filename.startswith('s3:'):
-        pkl_filename = S3Path.from_string(pkl_filename)
     if reconvert:
         # Organize by statement
         logger.info("Organizing by statement...")
@@ -185,12 +131,9 @@ def make_dataframe(reconvert, db_content, pkl_filename=None):
                 except KeyError:
                     nkey_errors += 1
                     error_keys.append((hash, num))
-                    if nkey_errors < 11:
+                    if nkey_errors < 10:
                         logger.warning('Missing key in agent name dict: '
                                        '(%s, %s)' % (hash, num))
-                    elif nkey_errors == 11:
-                        logger.warning('Got more than 10 key warnings: '
-                                       'muting further warnings.')
                     continue
 
             # Need at least two agents.
@@ -230,90 +173,52 @@ def make_dataframe(reconvert, db_content, pkl_filename=None):
         df = pd.DataFrame.from_dict(rows)
 
         if pkl_filename:
-            if isinstance(pkl_filename, S3Path):
-                upload_pickle_to_s3(obj=df, s3_path=pkl_filename)
-            else:
-                with open(pkl_filename, 'wb') as f:
-                    pickle.dump(df, f)
+            with open(pkl_filename, 'wb') as f:
+                pickle.dump(df, f)
     else:
         if not pkl_filename:
             logger.error('Have to provide pickle file if not reconverting')
             raise FileExistsError
         else:
-            if isinstance(pkl_filename, S3Path):
-                df = load_pickle_from_s3(pkl_filename)
-            else:
-                with open(pkl_filename, 'rb') as f:
-                    df = pickle.load(f)
+            with open(pkl_filename, 'rb') as f:
+                df = pickle.load(f)
     return df
 
 
-def get_parser():
+if __name__ == '__main__':
     parser = argparse.ArgumentParser(description='DB sif dumper',
-                                     usage=('Usage: dump_sif.py <db_dump_pkl> '
-                                            '<dataframe_pkl> <csv_file>'))
+        usage='Usage: dump_sif.py <db_dump_pkl> <dataframe_pkl> <csv_file>')
     parser.add_argument('--db-dump',
-                        help='A pickle of the database dump. If provided '
-                             'with --reload, this is the name of a new '
-                             'db dump pickle, otherwise this is assumed to '
-                             'be a cached pickle that already exists.')
+                        help='Dump a pickle of the database dump')
     parser.add_argument('--reload',
-                        help='Reload the database content from the database.',
+                        help='Reload from the database',
                         action='store_true',
-                        default=False)
+                        default=True)
     parser.add_argument('--dataframe',
-                        help='A pickle of the database dump processed '
-                             'into a pandas dataframe with pair '
-                             'interactions. If provided with the --reconvert '
-                             'option, this is the name of a new dataframe '
-                             'pickle, otherwise this is assumed to '
-                             'be a cached pickle that already exists.')
+                        help='Dump a pickle of the database dump processed '
+                             'into a pandas dataframe with pair interactions')
     parser.add_argument('--reconvert',
                         help='Re-run the dataframe processing on the db-dump',
                         action='store_true',
-                        default=False)
+                        default=True)
     parser.add_argument('--csv-file',
                         help='Dump a csv file with statistics of the database '
                              'dump')
-    parser.add_argument('--src-counts',
-                        help='If provided, also run and dump a pickled '
-                             'dictionary of the stratified evidence count '
-                             'per statement from each of the different '
-                             'sources.')
-    parser.add_argument('--s3',
-                        action='store_true',
-                        default=False,
-                        help='Upload files to the bigmech s3 bucket instead '
-                             'of saving them on the local disk.')
-    parser.add_argument('--s3-ymd',
-                        default=datetime.utcnow().strftime('%Y-%m-%d'),
-                        help='Set the dump sub-directory name on s3 '
-                             'specifying the date when the file was '
-                             'processed. Default: %Y-%m-%d of '
-                             'datetime.datetime.utcnow()')
-    parser.add_argument('--principal',
-                        action='store_true',
-                        default=False,
-                        help='Use the principal db instead of the readonly')
-    return parser
+    args = parser.parse_args()
 
-
-def dump_sif(df_file=None, db_res_file=None, csv_file=None, src_count_file=None,
-             reload=False, reconvert=False, ro=None):
-    if ro is None:
-        ro = get_db('primary')
-
+    dump_file = args.db_dump
+    df_file = args.dataframe
+    reload = args.reload
+    csv_file = args.csv_file
+    
     # Get the db content from a new DB dump or from file
     db_content = load_db_content(reload=reload, ns_list=NS_LIST,
-                                 pkl_filename=db_res_file, ro=ro)
-
+                                 pkl_filename=dump_file)
     # Convert the database query result into a set of pairwise relationships
-    df = make_dataframe(pkl_filename=df_file, reconvert=reconvert,
+    df = make_dataframe(pkl_filename=df_file, reconvert=True,
                         db_content=db_content)
 
     if csv_file:
-        if isinstance(csv_file, str) and csv_file.startswith('s3:'):
-            csv_file = S3Path.from_string(csv_file)
         # Aggregate rows by genes and stmt type
         logger.info("Saving to CSV...")
         filt_df = df.filter(items=['agA_ns', 'agA_id', 'agA_name',
@@ -322,70 +227,4 @@ def dump_sif(df_file=None, db_res_file=None, csv_file=None, src_count_file=None,
         type_counts = filt_df.groupby(by=['agA_ns', 'agA_id', 'agA_name',
                                           'agB_ns', 'agB_id', 'agB_name',
                                           'stmt_type']).sum()
-        # This requires package s3fs under the hood. See:
-        # https://pandas.pydata.org/pandas-docs/stable/whatsnew/v0.20.0.html#s3-file-handling
-        if isinstance(csv_file, S3Path):
-            try:
-                type_counts.to_csv(csv_file.to_string())
-            except Exception as e:
-                try:
-                    logger.warning('Failed to upload csv to s3 using direct '
-                                   's3 url, trying boto3: %s.' % e)
-                    s3 = get_s3_client(unsigned=False)
-                    csv_buf = StringIO()
-                    type_counts.to_csv(csv_buf)
-                    s3.put_object(Body=csv_buf.getvalue(), **csv_file.kw())
-                    logger.info('Uploaded CSV file to s3')
-                except Exception as second_e:
-                    logger.error('Failed to upload csv file with fallback '
-                                 'method')
-                    logger.exception(second_e)
-        # save locally
-        else:
-            type_counts.to_csv(csv_file)
-
-    if src_count_file:
-        _ = get_source_counts(src_count_file, ro=ro)
-    return
-
-
-def main():
-    args = get_parser().parse_args()
-
-    ymd = args.s3_ymd
-    if args.s3:
-        logger.info('Uploading to %s/%s/%s on s3 instead of saving locally'
-                    % (S3_SIF_BUCKET, S3_SUBDIR, ymd))
-    db_res_file = _pseudo_key(args.db_dump, ymd) if args.s3 and args.db_dump\
-        else args.db_dump
-    df_file = _pseudo_key(args.dataframe, ymd) if args.s3 and args.dataframe\
-        else args.dataframe
-    csv_file = _pseudo_key(args.csv_file, ymd) if args.s3 and args.csv_file\
-        else args.csv_file
-    src_count_file = _pseudo_key(args.src_counts, ymd) if args.s3 and \
-        args.src_counts else args.src_counts
-
-    reload = args.reload
-    if reload:
-        logger.info('Reloading the database content from the database')
-    else:
-        logger.info('Loading cached database content from %s' % db_res_file)
-
-    reconvert = args.reconvert
-    if reconvert:
-        logger.info('Reconverting database content into pandas dataframe')
-    else:
-        logger.info('Loading cached dataframe from %s' % df_file)
-
-    for f in [db_res_file, df_file, csv_file, src_count_file]:
-        if f:
-            logger.info('Using file name %s' % f)
-        else:
-            continue
-
-    dump_sif(df_file, db_res_file, csv_file, src_count_file, reload, reconvert,
-             get_db('primary') if args.principal else get_ro('primary'))
-
-
-if __name__ == '__main__':
-    main()
+        type_counts.to_csv(csv_file)
